Gen-AI/mcqgen/requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
langchain-community
langchain_openai

-e .